LibJava+java: Introduce the VM, and the `java` executable to run class files
This VM can execute a large number of instructions, but those only
operate on primitive types. No `wide` instruction support. No reference
support! Lot's of work to be done, but this can still execute a ton of
Java code :^)

The `java` program takes a path to a class file, and the name of a
`public static` method with no parameters.