LibJava: Introduce `StaticData` in the VM, keeping track in `m_static_data` per class file. Support the `getstatic` and `putstatic` opcode.
We now initialize the class by going through all the static fields, and
filling them with their `ConstantValue` attribute, or their default
value (often `0`). Then, if we find a `<clinit>` method, we call that,
which (should) further initialize all the static fields. Ofcourse, we
still assume all static fields are from the class we constructed with,
which is totally wrong!